requirements.txt
Django==3.1.1
PyMySQL==0.10.0
djangorestframework==3.11.1
django-environ==0.4.5
django-cors-headers==3.5.0
django-allauth==0.42.0
django-rest-auth==0.9.5
djangorestframework-jwt==1.11.0
